feature/refactor add Dockerfile draft, transport draft
