+++ Dockerfile
@@ -1,0 +1,9 @@
+FROM alpine:3.14
+
+WORKDIR /service
+
+COPY --chown=bot_operator weather-or-not-bot /service/.
+
+USER bot_operator
+
+ENTRYPOINT ["/service/weather-or-not-bot"]
