+++ internal/types/errors.go
@@ -1,3 +1,7 @@
 package types
 
-const ()
+import "github.com/pkg/errors"
+
+var (
+	Err = errors.New("some new error")
+)
